server/requirements.txt
fastapi==0.135.1
python-socketio==5.16.1
uvicorn[standard]==0.41.0
claude-code-sdk==0.0.25
